refactor: move all but controller to pkg
