+++ internal/controller/awsvalidator_controller.go
@@ -35,13 +35,13 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/validator-labs/validator-plugin-aws/api/v1alpha1"
-	"github.com/validator-labs/validator-plugin-aws/internal/aws"
-	"github.com/validator-labs/validator-plugin-aws/internal/constants"
-	"github.com/validator-labs/validator-plugin-aws/internal/validators"
-	"github.com/validator-labs/validator-plugin-aws/internal/validators/ami"
-	"github.com/validator-labs/validator-plugin-aws/internal/validators/iam"
-	"github.com/validator-labs/validator-plugin-aws/internal/validators/servicequota"
-	"github.com/validator-labs/validator-plugin-aws/internal/validators/tag"
+	"github.com/validator-labs/validator-plugin-aws/pkg/ami"
+	"github.com/validator-labs/validator-plugin-aws/pkg/aws"
+	"github.com/validator-labs/validator-plugin-aws/pkg/constants"
+	"github.com/validator-labs/validator-plugin-aws/pkg/iam"
+	"github.com/validator-labs/validator-plugin-aws/pkg/servicequota"
+	"github.com/validator-labs/validator-plugin-aws/pkg/tag"
+	validators "github.com/validator-labs/validator-plugin-aws/pkg/validate"
 	vapi "github.com/validator-labs/validator/api/v1alpha1"
 	"github.com/validator-labs/validator/pkg/types"
 	"github.com/validator-labs/validator/pkg/util"
