+++ pkg/ami/ami_validator_test.go
