+++ pkg/tag/tag_validator.go
@@ -15,8 +15,8 @@ import (
 	"github.com/validator-labs/validator/pkg/util"
 
 	"github.com/validator-labs/validator-plugin-aws/api/v1alpha1"
-	"github.com/validator-labs/validator-plugin-aws/internal/constants"
-	"github.com/validator-labs/validator-plugin-aws/internal/validators"
+	"github.com/validator-labs/validator-plugin-aws/pkg/constants"
+	"github.com/validator-labs/validator-plugin-aws/pkg/validate"
 )
 
 type tagAPI interface {
@@ -41,7 +41,7 @@ func NewTagRuleService(log logr.Logger, tagSvc tagAPI) *RuleService {
 func (s *RuleService) ReconcileTagRule(rule v1alpha1.TagRule) (*vapitypes.ValidationRuleResult, error) {
 
 	msg := fmt.Sprintf("All required %s tags were found", rule.ResourceType)
-	vr := validators.BuildValidationResult(rule.Name, msg, constants.ValidationTypeTag)
+	vr := validate.BuildValidationResult(rule.Name, msg, constants.ValidationTypeTag)
 
 	switch rule.ResourceType {
 	case "subnet":
